requirements.txt
tensorflow==2.12.0
numpy==1.22
Pillow==9.1.1
fastapi==0.78.0
uvicorn==0.17.6
python-multipart==0.0.6